# backend/src/main/resources/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "contact": {
      "email": "info@springonly.com",
      "name": "Equipo de Desarrollo de Spring Only"
    },
    "description": "API para información y gestión de créditos simples.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title": "Credit API REST",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://backend.springonly.com",
      "description": "Servidor Spring Only"
    }
  ],
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
      "GetCustomerByIdResponse": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "customerTypeId": {
            "type": "string"
          },
          "riskCategoryId": {
            "type": "string"
          },
          "lineOfCreditAmount": {
            "type": "number"
          },
          "relationshipManagerId": {
            "type": "string"
          },
          "writtenAt": {
            "type": "string",
            "format": "date-time",
            "examples": [
              "2022-03-10T12:15:50-04:00"
            ]
          }
        }
      },
...
      "GetRelationshipManagerByIdResponse": {
        "type": "object",
        "properties": {
          "relationshipManagerId": {
            "type": "string"
          },
          "relationshipManagerName": {
            "type": "string"
          },
          "threadId": {
            "type": "string"
          },
          "writtenAt": {
            "type": "string",
            "format": "date-time",
            "examples": [
              "2022-03-10T12:15:50-04:00"
            ]
          }
        }
      },
...
      "LoginRelationshipManagerRequest": {
        "type": "object",
        "properties": {
          "relationshipManagerId": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginRelationshipManagerResponse": {
        "type": "object",
        "properties": {
          "relationshipManagerId": {
            "type": "string"
          },
          "relationshipManagerName": {
            "type": "string"
          },
          "threadId": {
            "type": "string"
          },
          "writtenAt": {
            "type": "string",
            "format": "date-time",
            "examples": [
              "2022-03-10T12:15:50-04:00"
            ]
          }
        }
      },
...
      "UpdateRelationshipManagerThreadIdRequest": {
        "type": "object",
        "properties": {
          "threadId": {
            "type": "string",
            "format": "uuid",
            "pattern": "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
          }
        }
      },
      "UpdateRelationshipManagerThreadIdResponse": {
        "type": "object",
        "properties": {
          "relationshipManagerId": {
            "type": "string"
          },
          "relationshipManagerName": {
            "type": "string"
          },
          "threadId": {
            "type": "string"
          },
          "writtenAt": {
            "type": "string",
            "format": "date-time",
            "examples": [
              "2022-03-10T12:15:50-04:00"
            ]
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/relationship-managers/login": {
      "parameters": [
        {
          "name": "X-RelationshipManager-Id",
          "in": "header",
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "summary": "Login relationship manager",
        "description": "Authenticate a relationship manager and return session or token details.",
        "operationId": "loginRelationshipManager",
        "requestBody": {
          "description": "LoginRelationshipManagerRequest payload",
          "content": {
            "application/json": {
              "examples": {
                "Login request": {
                  "summary": "Request to authenticate a relationship manager",
                  "value": {
                    "relationshipManagerId": "RM001",
                    "password": "s3cr3t"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/LoginRelationshipManagerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "examples": {
                  "Login successful": {
                    "summary": "Authenticated relationship manager response",
                    "value": {
                      "relationshipManagerId": "RM001",
                      "relationshipManagerName": "Ana Pérez",
                      "threadId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                      "writtenAt": "2025-10-07T22:27:36-05:00"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/LoginRelationshipManagerResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "examples": {
                  "Unauthorized login": {
                    "summary": "Invalid credentials example",
                    "value": {
                      "message": "Usuario y/o password incorrectos",
                      "code": "RM001"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "tags": [
          "Relationship Manager Resource"
        ]
      }
    },
    "/api/relationship-managers/{relationshipManagerId}": {
      "parameters": [
        {
          "name": "X-RelationshipManager-Id",
          "in": "header",
          "schema": {
            "type": "string"
          }
        }
      ],
      "patch": {
        "summary": "Update thread id",
        "description": "Update the thread id information for a relationship manager.",
        "operationId": "updateRelationshipManagerThreadId",
        "parameters": [
          {
            "name": "relationshipManagerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "UpdateRelationshipManagerThreadIdRequest payload",
          "content": {
            "application/json": {
              "examples": {
                "Update thread id request": {
                  "summary": "Request to update relationship manager thread id",
                  "value": {
                    "threadId": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/UpdateRelationshipManagerThreadIdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Relationship manager thread id updated",
            "content": {
              "application/json": {
                "examples": {
                  "Update thread id success": {
                    "summary": "Updated relationship manager thread id response",
                    "value": {
                      "relationshipManagerId": "RM001",
                      "relationshipManagerName": "Ana Pérez",
                      "threadId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                      "writtenAt": "2025-10-07T22:27:36-05:00"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/UpdateRelationshipManagerThreadIdResponse"
                }
              }
            }
          },
          "404": {
            "description": "Relationship manager not found",
            "content": {
              "application/json": {
                "examples": {
                  "Update thread id not found": {
                    "summary": "Relationship manager not found example",
                    "value": {
                      "message": "El Ejecutivo de Cuenta indicado no existe",
                      "code": "RM002"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Relationship Manager Resource"
        ]
      },
      "get": {
        "summary": "Get relationship manager",
        "description": "Retrieve relationship manager details by id.",
        "operationId": "getRelationshipManagerById",
        "parameters": [
          {
            "name": "relationshipManagerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Relationship manager retrieved successfully",
            "content": {
              "application/json": {
                "examples": {
                  "GetRelationshipManagerByIdResponse example": {
                    "summary": "A relationship manager retrieved by id",
                    "value": {
                      "relationshipManagerId": "RM001",
                      "relationshipManagerName": "Ana Pérez",
                      "threadId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                      "writtenAt": "2025-10-07T22:27:36-05:00"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/GetRelationshipManagerByIdResponse"
                }
              }
            }
          },
          "404": {
            "description": "Relationship manager not found",
            "content": {
              "application/json": {
                "examples": {
                  "GetRelationshipManager not found": {
                    "summary": "Requested relationship manager does not exist",
                    "value": {
                      "message": "El Ejecutivo de Cuenta indicado no existe",
                      "code": "RM003"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Relationship Manager Resource"
        ]
      }
    },
    "/api/relationship-managers/{relationshipManagerId}/customers": {
      "parameters": [
        {
          "name": "X-RelationshipManager-Id",
          "in": "header",
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "List customers",
        "description": "List customers assigned to a specific relationship manager.",
        "operationId": "listCustomersByRelationshipManagerById",
        "parameters": [
          {
            "name": "relationshipManagerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customers list for relationship manager",
            "content": {
              "application/json": {
                "examples": {
                  "Customers list": {
                    "summary": "A list of customers for a relationship manager",
                    "value": [
                      {
                        "customerId": "CUST12345",
                        "customerName": "John Doe",
                        "customerTypeId": "MICRO EMPRESA",
                        "riskCategoryId": "DU",
                        "lineOfCreditAmount": 50000.0,
                        "relationshipManagerId": "RM001",
                        "writtenAt": "2024-12-14T19:37:26-05:00"
                      },
                      {
                        "customerId": "CUST67890",
                        "customerName": "María López",
                        "customerTypeId": "PERSONAL",
                        "riskCategoryId": "PP",
                        "lineOfCreditAmount": 15000.0,
                        "relationshipManagerId": "RM001",
                        "writtenAt": "2025-01-20T15:00:00-05:00"
                      }
                    ]
                  }
                },
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetCustomerByIdResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No customers found for relationship manager",
            "content": {
              "application/json": {
                "examples": {
                  "No customers found": {
                    "summary": "No customers for given relationship manager",
                    "value": {
                      "message": "El Ejecutivo de Cuentas no tiene clientes asignados",
                      "code": "RM002"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Relationship Manager Resource"
        ]
      }
    }
  }
}